requirements.txt
anyio==3.6.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cx-Oracle==8.3.0
fastapi==0.83.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
zipp==3.8.1
